feat(admin): improve video review with file isolation and cleanup

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/po/UserVideoActionArchive.java
@@ -33,7 +33,7 @@ public class UserVideoActionArchive
     /**
      * 1:视频点赞 2:视频收藏 3:视频投币
      */
-    private Integer actionType;
+    private Short actionType;
 
     /**
      * 投币数量
